+++ tests/helpers/Ratchet/Mock/WampComponent.php
@@ -3,6 +3,7 @@ namespace Ratchet\Mock;
 use Ratchet\Wamp\WampServerInterface;
 use Ratchet\WebSocket\WsServerInterface;
 use Ratchet\ConnectionInterface;
+use Throwable;
 
 class WampComponent implements WampServerInterface, WsServerInterface {
     public $last = array();
@@ -37,7 +38,7 @@ class WampComponent implements WampServerInterface, WsServerInterface {
         $this->last[__FUNCTION__] = func_get_args();
     }
 
-    public function onError(ConnectionInterface $conn, \Exception $e) {
+    public function onError(ConnectionInterface $conn, Throwable $e) {
         $this->last[__FUNCTION__] = func_get_args();
     }
 }