+++ tests/helpers/Ratchet/NullComponent.php
@@ -1,9 +1,8 @@
 <?php
 namespace Ratchet;
-use Ratchet\ConnectionInterface;
-use Ratchet\MessageComponentInterface;
 use Ratchet\WebSocket\WsServerInterface;
 use Ratchet\Wamp\WampServerInterface;
+use Throwable;
 
 class NullComponent implements MessageComponentInterface, WsServerInterface, WampServerInterface {
     public function onOpen(ConnectionInterface $conn) {}
@@ -12,7 +11,7 @@ class NullComponent implements MessageComponentInterface, WsServerInterface, Wam
 
     public function onClose(ConnectionInterface $conn) {}
 
-    public function onError(ConnectionInterface $conn, \Exception $e) {}
+    public function onError(ConnectionInterface $conn, Throwable $e) {}
 
     public function onCall(ConnectionInterface $conn, $id, $topic, array $params) {}