+++ tests/unit/Server/IoConnectionTest.php
@@ -1,17 +1,20 @@
 <?php
 namespace Ratchet\Application\Server;
+use PHPUnit\Framework\TestCase;
 use Ratchet\Server\IoConnection;
 
 /**
  * @covers Ratchet\Server\IoConnection
  */
-class IoConnectionTest extends \PHPUnit_Framework_TestCase {
+class IoConnectionTest extends TestCase {
     protected $sock;
     protected $conn;
 
-    public function setUp() {
-        $this->sock = $this->getMock('\\React\\Socket\\ConnectionInterface');
+    public function setUp() : void {
+        $this->sock = $this->createMock('\\React\\Socket\\ConnectionInterface');
         $this->conn = new IoConnection($this->sock);
+
+        parent::setUp();
     }
 
     public function testCloseBubbles() {