+++ tests/unit/Session/Serialize/PhpHandlerTest.php
@@ -1,15 +1,17 @@
 <?php
 namespace Ratchet\Session\Serialize;
-use Ratchet\Session\Serialize\PhpHandler;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers Ratchet\Session\Serialize\PhpHandler
  */
-class PhpHandlerTest extends \PHPUnit_Framework_TestCase {
+class PhpHandlerTest extends TestCase {
     protected $_handler;
 
-    public function setUp() {
+    public function setUp() : void {
         $this->_handler = new PhpHandler;
+
+        parent::setUp();
     }
 
     public function serializedProvider() {