+++ tests/unit/Session/Storage/VirtualSessionStoragePDOTest.php
@@ -1,11 +1,13 @@
 <?php
 namespace Ratchet\Session\Storage;
+use PHPUnit\Framework\TestCase;
 use Ratchet\Session\Serialize\PhpHandler;
+use Ratchet\Tests\Session\InMemoryOptionsHandler;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBag;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBag;
 use Symfony\Component\HttpFoundation\Session\Storage\Handler\PdoSessionHandler;
 
-class VirtualSessionStoragePDOTest extends \PHPUnit_Framework_TestCase {
+class VirtualSessionStoragePDOTest extends TestCase {
     /**
      * @var VirtualSessionStorage
      */
@@ -13,9 +15,11 @@ class VirtualSessionStoragePDOTest extends \PHPUnit_Framework_TestCase {
 
     protected $_pathToDB;
 
-    public function setUp() {
+    public function setUp() : void {
         if (!extension_loaded('PDO') || !extension_loaded('pdo_sqlite')) {
-            return $this->markTestSkipped('Session test requires PDO and pdo_sqlite');
+            $this->markTestSkipped('Session test requires PDO and pdo_sqlite');
+
+            return;
         }
 
         $schema = <<<SQL
@@ -36,13 +40,17 @@ SQL;
 
         $sessionHandler = new PdoSessionHandler($dsn);
         $serializer = new PhpHandler();
-        $this->_virtualSessionStorage = new VirtualSessionStorage($sessionHandler, 'foobar', $serializer);
+        $this->_virtualSessionStorage = new VirtualSessionStorage($sessionHandler, 'foobar', $serializer, new InMemoryOptionsHandler());
         $this->_virtualSessionStorage->registerBag(new FlashBag());
         $this->_virtualSessionStorage->registerBag(new AttributeBag());
+
+        parent::setUp();
     }
 
-    public function tearDown() {
+    public function tearDown() : void {
         unlink($this->_pathToDB);
+
+        parent::tearDown();
     }
 
     public function testStartWithDSN() {